requirements.txt
absl-py==2.1.0
accelerate==0.30.1
addict==2.4.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
argparse==1.4.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
automat==22.10.0
av==11.0.0
basicsr==1.4.2
beautifulsoup4==4.12.3
bezier==2023.7.28
buildtools==1.0.6
causal-conv1d==1.1.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
cycler==0.12.1
cython==3.0.10
dataproperty==1.0.1
datasets==2.19.1
decord==0.6.0
deepspeed==0.13.1
dill==0.3.8
docker-pycreds==0.4.0
docopt==0.6.2
einops==0.7.0
evaluate==0.4.2
exceptiongroup==1.2.1
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.3
furl==2.1.3
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.5.4
gdown==5.2.0
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.0
h5py==3.11.0
hjson==3.1.0
huggingface-hub==0.23.2
hyperlink==21.0.0
imageio==2.33.1
importlib-metadata==7.2.1
incremental==22.10.0
iniconfig==2.0.0
iopath==0.1.10
joblib==1.4.2
jsonlines==4.0.0
keras==3.3.3
kiwisolver==1.4.5
lazy-loader==0.4
libclang==18.1.1
lm-eval==0.4.1
lmdb==1.4.1
lxml==5.2.2
mamba-ssm==1.1.1
markdown==3.6
markdown-it-py==3.0.0
markupsafe==2.1.5
matplotlib==3.9.0
mbstrdecoder==1.1.3
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.0.5
multiprocess==0.70.16
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.7.0.68
opt-einsum==3.3.0
optree==0.11.0
orderedmultidict==1.0.1
packaging==24.0
pandas==2.2.1
pathvalidate==3.2.0
peft==0.11.1
pillow==10.1.0
platformdirs==4.2.2
pluggy==1.5.0
portalocker==2.8.2
progressbar2==4.4.2
protobuf==4.25.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pybind11==2.12.0
pycocotools==2.0.8
pydantic==2.7.2
pydantic-core==2.18.3
pygments==2.18.0
pynvml==11.5.0
pyparsing==3.1.2
pyqt5==5.15.10
pyqt5-qt5==5.15.2
pyqt5-sip==12.13.0
pytablewriter==1.2.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-utils==3.8.2
pyyaml==6.0.1
pytz==2024.1
redo==2.0.4
regex==2023.10.3
requests==2.31.0
rich==13.7.1
rouge-score==0.1.2
sacrebleu==2.4.2
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.12.0
sentry-sdk==2.3.1
setproctitle==1.3.3
setuptools==68.2.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
sqlalchemy==2.0.30
sqlitedict==2.1.0
submitit==1.5.1
sympy==1.12.1
tabledata==1.3.3
tabulate==0.9.0
tb-nightly==2.18.0a20240625
tcolorpy==0.1.6
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorboardx==2.6.2.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==0.4.12
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
tqdm-multiprocess==0.0.11
transformers==4.36.1
triton==2.2.0
twisted==24.3.0
typepy==1.3.2
typing-extensions==4.12.0
tzdata==2024.1
wandb==0.16.2
wcwidth==0.2.13
werkzeug==3.0.3
wheel==0.42.0
wrapt==1.16.0
xformers==0.0.24
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
zope-interface==6.4.post2
zstandard==0.22.0
